+++ sql/expression/function/aggregation/window_functions.go
@@ -15,7 +15,7 @@
 package aggregation
 
 import (
-	sort2 "github.com/dolthub/go-mysql-server/sql/sort"
+	sort2 "github.com/dolthub/go-mysql-server/sql/sorters"
 	"math"
 	"sort"
 	"strings"
