+++ sql/iters/rel_iters.go
@@ -16,7 +16,7 @@ package iters
 
 import (
 	"fmt"
-	sort2 "github.com/dolthub/go-mysql-server/sql/sort"
+	sort2 "github.com/dolthub/go-mysql-server/sql/sorters"
 	"io"
 	"sort"
